requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.84.0
anyio==4.12.1
async-timeout==4.0.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.1.8
contourpy==1.3.0
cryptography==46.0.5
cycler==0.12.1
distro==1.9.0
docstring_parser==0.17.0
exceptiongroup==1.3.1
fastapi==0.128.8
filetype==1.2.0
fonttools==4.60.2
google-ai-generativelanguage==0.10.0
google-api-core==2.30.0
google-auth==2.49.1
googleapis-common-protos==1.73.0
grpcio==1.78.0
grpcio-status==1.78.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
importlib_resources==6.5.2
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.7
langchain==0.3.28
langchain-anthropic==0.3.22
langchain-core==0.3.83
langchain-google-genai==2.1.12
langchain-openai==0.3.35
langchain-text-splitters==0.3.11
langgraph==0.6.11
langgraph-checkpoint==2.1.2
langgraph-prebuilt==0.6.5
langgraph-sdk==0.2.9
langsmith==0.4.37
matplotlib==3.9.4
numpy==2.0.2
openai==2.26.0
orjson==3.11.5
ormsgpack==1.11.0
packaging==25.0
pandas==2.3.3
pillow==11.3.0
proto-plus==1.27.1
protobuf==6.33.5
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.1.15
requests==2.32.5
requests-toolbelt==1.0.0
scikit-learn==1.6.1
scipy==1.13.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.48
starlette==0.49.3
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.39.0
xgboost==2.1.4
xxhash==3.6.0
zipp==3.23.0
zstandard==0.25.0
